+++ lib/heroku_tool/commander.rb
@@ -4,9 +4,12 @@ module HerokuTool
   class Commander
     # @return [HerokuTool::HerokuTargets::HerokuTarget]
     attr_reader :target
-    def initialize(target, **options)
+    attr_reader :configuration
+
+    def initialize(target, configuration:, **options)
       @target = target
       @migrate_outside_of_release_phase = target&.migrate_in_release_phase ? false : options[:migrate]
+      @configuration = configuration
     end
 
     def deploy(deploy_ref, with_maintenance:)
@@ -14,7 +17,7 @@ module HerokuTool
       puts "Deploy #{deploy_ref_description} to #{target} with migrate=#{target.migrate_in_release_phase ? "(during release phase)" : migrate_outside_of_release_phase?} with_maintenance=#{with_maintenance} "
 
       output_to_be_deployed(deploy_ref)
-      Heroku::Configuration.before_deploying(self, target, deploy_ref_description)
+      configuration.before_deploying(self, target, deploy_ref_description)
       successful_push = puts_and_system "git push -f #{target.git_remote} #{deploy_ref || target}^{}:#{target.heroku_target_ref}"
 
       return false unless successful_push
@@ -24,25 +27,25 @@ module HerokuTool
         puts_and_system "heroku run rake db:migrate -a #{target.heroku_app}"
       end
 
-      app_revision_env_var = Heroku::Configuration.app_revision_env_var
+      app_revision_env_var = configuration.app_revision_env_var
       if app_revision_env_var && app_revision_env_var != "HEROKU_SLUG_COMMIT"
         # HEROKU_SLUG_COMMIT is automatically set by https://devcenter.heroku.com/articles/dyno-metadata
         puts_and_system %{heroku config:set #{app_revision_env_var}=$(git describe --always #{deploy_ref}) -a #{target.heroku_app}}
       end
 
       maintenance_off if with_maintenance
-      Heroku::Configuration.after_deploying(self, target, deploy_ref_description)
+      configuration.after_deploying(self, target, deploy_ref_description)
       true
     end
 
     def maintenance_on
       puts_and_system "heroku maintenance:on -a #{target.heroku_app}"
-      puts_and_system "heroku config:set #{Heroku::Configuration.maintenance_mode_env_var}=true -a #{target.heroku_app}"
+      puts_and_system "heroku config:set #{configuration.maintenance_mode_env_var}=true -a #{target.heroku_app}"
     end
 
     def maintenance_off
       puts_and_system "heroku maintenance:off -a #{target.heroku_app}"
-      puts_and_system "heroku config:unset #{Heroku::Configuration.maintenance_mode_env_var} -a #{target.heroku_app}"
+      puts_and_system "heroku config:unset #{configuration.maintenance_mode_env_var} -a #{target.heroku_app}"
     end
 
     def migrate_outside_of_release_phase?
@@ -57,7 +60,7 @@ module HerokuTool
       puts "------------------------------"
       puts " Deploy to #{target}:"
       puts "------------------------------"
-      system_with_clean_env "git --no-pager log $(heroku config:get #{Heroku::Configuration.app_revision_env_var} -a #{target.heroku_app})..#{since_deploy_ref || target.deploy_ref}"
+      system_with_clean_env "git --no-pager log $(heroku config:get #{configuration.app_revision_env_var} -a #{target.heroku_app})..#{since_deploy_ref || target.deploy_ref}"
       puts "------------------------------"
     end