+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v0.9.0
 
+- BREAKING CHANGE: Heroku::Configuration has been renamed to HerokuTool::Configuration and is now must be an instance not a class method
 - BREAKING CHANGE: drop set_message for simplicity - you can do that as part of Configuration.before_deploying and Configuration.after_deploying
 - don't continue if git push fails (output AND return status and don't do migration, deployment tracking etc)